requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
openai
langchain
faiss-cpu
python-dotenv
